refactor(Utils): 패키지 속 클래스/메소드 정리
