+++ README.md
@@ -163,7 +163,11 @@
 - [x] UI(view) 로직과 비즈니스 로직 분리
 - [x] .을 1개까지만 허용
 - [x] while문 재귀로 변경
+- [ ] 클래스가 한가지 일만 하도록 하기
 - [ ] 메소드가 한가지 일만 하도록 하기
+- [ ] 더 많은 클래스의 기능을 사용하는 곳으로 메소드 이동
+- [ ] 조건을 메소드 대신 변수로 표현
+- [ ] 필요없는 setter는 제거
 
 ---
