+++ src/main/java/vendingmachine/Model/Money.java
@@ -1,5 +1,6 @@
 package vendingmachine.Model;
 
+import vendingmachine.Utils.Converter;
 import vendingmachine.Utils.Validator.MoneyValidator;
 import vendingmachine.Utils.Validator.UserMoneyValidator;
