+++ src/main/java/vendingmachine/Utils/Util.java
@@ -2,13 +2,15 @@ package vendingmachine.Utils;
 
 import java.util.Arrays;
 
+import camp.nextstep.edu.missionutils.Console;
 import camp.nextstep.edu.missionutils.Randoms;
-import vendingmachine.Model.Coin;
 
 public class Util {
-	public static Coin randomCoin() {
-		return Coin.getCoin(
-			Randoms.pickNumberInList(Arrays.asList(Constants.COIN_NUMS))
-		);
+	public static int randomCoinNumber() {
+		return Randoms.pickNumberInList(Arrays.asList(Constants.COIN_NUMBERS));
+	}
+
+	public static String getInput() {
+		return Console.readLine();
 	}
 }
